requirements.txt
# FastAPI Framework y dependencias core
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Base de datos y ORM
sqlmodel
sqlalchemy

# Validación y utilidades
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]

# Testing
pytest
pytest-asyncio
httpx
requests
ruff
black
isort
pre-commit
